Implemented Synthesizer::renderDepth
NOTE: Failing test case indicates ERP triangles need to be split and culled. I prefer to work on the other Synthesizer methods first.

+++ Renderer/include/TMIV/Renderer/Synthesizer.h
@@ -40,7 +40,7 @@ namespace TMIV::Renderer {
 class Synthesizer : public ISynthesizer {
 public:
   Synthesizer(const Common::Json &node);
-  Synthesizer(double rayAngleParam, double depthParam, double stretchingParam);
+  Synthesizer(float rayAngleParam, float depthParam, float stretchingParam);
   Synthesizer(const Synthesizer &) = delete;
   Synthesizer(Synthesizer &&) = default;
   Synthesizer &operator=(const Synthesizer &) = delete;
